23/11: Sửa cấu trúc code
